double data type issue fixed

+++ app/src/main/java/com/retailetics/ezycollect/data/remote/dto/CartDetailResponse.kt
@@ -3,12 +3,12 @@ package com.retailetics.ezycollect.data.remote.dto
 
 data class ShoppingCartDetails(
     val items: List<Item>,
-    val subtotal: Int
+    val subtotal: Double
 )
 
 data class Item(
     val id: Int,
-    val price: Int,
+    val price: Double,
     val product_name: String,
     val quantity: Int
 )
